Update GPTflix to become a QA retrieval tutorial.
