+++ chat/.env
@@ -1,1 +1,1 @@
-OPENAI_API_KEY=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
+OPENAI_API_KEY=
